+++ raaghu-mfe/projects/demo-ui/src/app/app-routing.module.ts
@@ -1,24 +1,16 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AppComponent } from './app.component';
-import { DemoUiComponent } from './demo-ui/demo-ui.component';
 
 const routes: Routes = [
   {
     path: '',
     pathMatch: 'full',
     component: AppComponent,
-    children: [
-      {
-        path: '',
-        component: DemoUiComponent,
-        pathMatch: 'full'
-      }
-    ]
   }
 ];
 @NgModule({
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
 })
-export class AppRoutingModule {}
+export class AppRoutingModule { }
