+++ raaghu-mfe/projects/demo-ui/src/app/app.component.html
@@ -1,2 +1,3 @@
-
-<router-outlet></router-outlet>
+<div @fadeAnimation>
+  <mfe-loader [config]="rdsDemouiMfeConfig"></mfe-loader>
+</div>
