+++ raaghu-mfe/projects/demo-ui/src/app/app.component.ts
@@ -2,14 +2,45 @@ import { Component, OnInit } from '@angular/core';
 import { selectDefaultLanguage } from '@libs/state-management';
 import { Store } from '@ngrx/store';
 import { TranslateService } from '@ngx-translate/core';
+import { ComponentLoaderOptions } from '@libs/shared';
+import { transition, trigger, query, style, animate, } from '@angular/animations';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+  styleUrls: ['./app.component.scss'],
+  animations: [
+    trigger('fadeAnimation', [
+      transition('* <=> *', [
+        query(':enter',
+          [
+            style({ opacity: 0 })
+          ],
+          { optional: true }
+        ),
+        query(':leave',
+          [
+            style({ opacity: 1 }),
+            animate('1s', style({ opacity: 0 }))
+          ],
+          { optional: true }
+        ),
+        query(':enter',
+          [
+            style({ opacity: 0 }),
+            animate('0.6s', style({ opacity: 1 }))
+          ],
+          { optional: true }
+        )
+      ])
+    ])
+  ]
 })
 export class AppComponent implements OnInit {
   title = 'demo-ui';
+  rdsDemouiMfeConfig: ComponentLoaderOptions = {
+    name: 'RdsCompDemoui',
+  };
   constructor(private store: Store, private translate: TranslateService) { }
   ngOnInit(): void {
     this.store.select(selectDefaultLanguage).subscribe((res: any) => {
