+++ raaghu-mfe/projects/demo-ui/src/app/app.module.ts
@@ -1,21 +1,18 @@
 import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { DemoUiComponent } from './demo-ui/demo-ui.component';
-
 import { NgxTranslateModule, SharedModule } from '@libs/shared';
+import { CommonModule } from '@angular/common';
 
 @NgModule({
   declarations: [
     AppComponent,
-    DemoUiComponent
   ],
   imports: [    
     AppRoutingModule,
     SharedModule,
     NgxTranslateModule.forRoot(),
-    
+    CommonModule
   ],
   providers: [],
   bootstrap: [AppComponent]
