--- raaghu-mfe/projects/demo-ui/src/app/demo-ui/demo-ui.component.html
@@ -1,3 +1,0 @@
-<div [@fadeAnimation]="isAnimation ? 'fadeAnimation' : ''">
-  <mfe-loader [config]="rdsDemouiMfeConfig"></mfe-loader>
-</div>
