+++ raaghu-mfe/projects/maintenance/src/app/app-routing.module.ts
@@ -1,26 +1,25 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AppComponent } from './app.component';
-import { MaintenanceComponent } from './maintenance/maintenance.component';
 
-// const routes: Routes = [
-//   {
-//     path: '',
-//     pathMatch: 'full',
-//     component: AppComponent,
-//   },
-// ];
-const routes: Routes = [
-  {
-    path: '',
-    redirectTo: 'maintenance'
-  },
-  {
-    path: 'maintenance',
-    component: MaintenanceComponent,
-   
-  },
+ const routes: Routes = [
+   {
+     path: '',
+     pathMatch: 'full',
+     component: AppComponent,
+   },
  ];
+//const routes: Routes = [
+//  {
+//    path: '',
+//  //  redirectTo: 'maintenance'
+//  },
+//  //{
+//  //  path: 'maintenance',
+//  //  component: MaintenanceComponent,
+   
+//  //},
+// ];
 @NgModule({
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
