+++ raaghu-mfe/projects/maintenance/src/app/app.module.ts
@@ -4,16 +4,15 @@ import { RdsBadgeModule, RdsButtonModule, RdsIconModule, RdsModalModule, RdsNavT
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { MaintenanceComponent } from './maintenance/maintenance.component';
-import { CacheComponent } from './cache/cache.component';
-import { WebsitelogComponent } from './websitelog/websitelog.component';
+
 import { NgxTranslateModule, SharedModule } from '@libs/shared';
 import { MaintenanceEffects, MaintenanceReducer, WebsitelogReducer } from '@libs/state-management';
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 import { RdsFabMenuModule } from 'projects/libs/rds-elements/src/rds-fab-menu/src/public-api';
-import { RdsMaintenanceShimmerComponent } from './maintenance/rds-maintenance-shimmer/rds-maintenance-shimmer.component';
+
 import { NgxShimmerLoadingModule } from 'ngx-shimmer-loading';
+import { CommonModule } from '@angular/common';
 export const MaintenanceReducersMap = {
   maintenances: MaintenanceReducer,
   Websitelogs: WebsitelogReducer,
@@ -21,15 +20,11 @@ export const MaintenanceReducersMap = {
 };
 @NgModule({
   declarations: [
-    AppComponent,
-    MaintenanceComponent,
-    CacheComponent,
-    WebsitelogComponent,
-    RdsMaintenanceShimmerComponent
+    AppComponent,    
 
   ],
   imports: [
-    BrowserModule,
+    CommonModule,
     AppRoutingModule,
     RdsNavTabModule,RdsModalModule,RdsButtonModule,
     RdsBadgeModule,SharedModule,
