--- raaghu-mfe/projects/maintenance/src/app/cache/cache.component.html
@@ -1,1 +1,0 @@
-<mfe-loader [config]="rdscacheMfeConfig"></mfe-loader>
