--- raaghu-mfe/projects/maintenance/src/app/maintenance/rds-maintenance-shimmer/rds-maintenance-shimmer.component.html
@@ -1,25 +1,0 @@
-<p>rds-maintenance-shimmer works!</p>
-
-<div class="row">
-  <div class="col-md-12 text-end mb-2">
-   <ngx-shimmer-loading [direction]="'ltr'" [shape]="'rect'" [width]="'10%'" [height]="'100%'"></ngx-shimmer-loading>
-  </div>
-  <div class="col-12">
-   <div class="card p-2 h-100 border-0 rounded-0 card-full-stretch">
-     <div col-md-2>
-       <ngx-shimmer-loading [direction]="'ltr'" [shape]="'rect'" [width]="'10%'" [height]="'150%'"></ngx-shimmer-loading>
-     </div>
-     <div class="mt-3">
-         <ngx-shimmer-loading [direction]="'ltr'" [shape]="'rect'" [width]="'100%'" [height]="'20%'"></ngx-shimmer-loading>
-         <ngx-shimmer-loading [direction]="'ltr'" [shape]="'rect'" [width]="'100%'" [height]="'20%'"></ngx-shimmer-loading>
-         <ngx-shimmer-loading [direction]="'ltr'" [shape]="'rect'" [width]="'100%'" [height]="'20%'"></ngx-shimmer-loading>
-         <ngx-shimmer-loading [direction]="'ltr'" [shape]="'rect'" [width]="'100%'" [height]="'20%'"></ngx-shimmer-loading>
-         <ngx-shimmer-loading [direction]="'ltr'" [shape]="'rect'" [width]="'100%'" [height]="'20%'"></ngx-shimmer-loading>
-         <ngx-shimmer-loading [direction]="'ltr'" [shape]="'rect'" [width]="'100%'" [height]="'20%'"></ngx-shimmer-loading>
-         <ngx-shimmer-loading [direction]="'ltr'" [shape]="'rect'" [width]="'100%'" [height]="'20%'"></ngx-shimmer-loading>
-         <ngx-shimmer-loading [direction]="'ltr'" [shape]="'rect'" [width]="'100%'" [height]="'20%'"></ngx-shimmer-loading>
-       </div>
-     </div>
-   </div>
- </div>
-
