--- raaghu-mfe/projects/maintenance/src/app/maintenance/rds-maintenance-shimmer/rds-maintenance-shimmer.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { RdsMaintenanceShimmerComponent } from './rds-maintenance-shimmer.component';
-
-describe('RdsMaintenanceShimmerComponent', () => {
-  let component: RdsMaintenanceShimmerComponent;
-  let fixture: ComponentFixture<RdsMaintenanceShimmerComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ RdsMaintenanceShimmerComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(RdsMaintenanceShimmerComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
