--- raaghu-mfe/projects/maintenance/src/app/maintenance/rds-maintenance-shimmer/rds-maintenance-shimmer.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-rds-maintenance-shimmer',
-  templateUrl: './rds-maintenance-shimmer.component.html',
-  styleUrls: ['./rds-maintenance-shimmer.component.scss']
-})
-export class RdsMaintenanceShimmerComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
