--- raaghu-mfe/projects/maintenance/src/app/websitelog/websitelog.component.html
@@ -1,2 +1,0 @@
-
-<mfe-loader [config]="rdswebsiteMfeConfig"></mfe-loader>
