+++ raaghu-mfe/projects/organization-unit/src/app/app.component.html
@@ -1,22 +1,23 @@
-<div [@fadeAnimation]="isAnimation ? 'fadeAnimation' : ''">
+<div>
   <div [class.card]="selectedTreeNode==0">
     <div class="row" [class.card-body]="selectedTreeNode==0">
       <div class="col-md-6 gutter-b">
-        <div class="card p-2 border-0 rounded-0 vh-88">
-          <div class="card-header bg-transparent">
-            <h5 class="card-title">{{'Organization Tree'| translate}}</h5>
-          </div>
-          <div >
+        <div >
+          <div class="card p-2 border-0 rounded-0 vh-88" [@fadeAnimation]="isAnimation ? 'fadeAnimation' : ''">
+
+            <div class="card-header bg-transparent">
+              <h5 class="card-title">{{'Organization Tree'| translate}}</h5>
+            </div>
             <div class="card-body">
               <mfe-loader [config]="rdsOrganizationTreeConfig"></mfe-loader>
             </div>
           </div>
-         
         </div>
+        
       </div>
 
       <div class="col-md-6 gutter-b">
-        <div class="card p-2 border-0 rounded-0 vh-88">
+        <div class="card p-2 border-0 rounded-0 vh-88" [@fadeAnimation]="isAnimation ? 'fadeAnimation' : ''">
           <!-- <div class="card-header bg-transparent" *ngIf="selectedTreeNode == 0">
             <h5 class="card-title">{{'Select an organization unit to see members' | translate}}</h5>
           </div> -->
@@ -48,16 +49,16 @@
                 <div naveContent class="row tab-content" id="nav-tabContent">
                   <div class="tab-pane fade show active" id="Members" role="tabpanel" aria-labelledby="nav-home-tab">
                     <div [@fadeAnimation]="isAnimation ? 'fadeAnimation' : ''"></div>
-                      <div class="row">
-                        <mfe-loader [config]="rdsDataTableForMemberMfeConfig"></mfe-loader>
-                      </div>
-                    </div>
-                    <div class="tab-pane fade" id="Roles" role="tabpanel" aria-labelledby="nav-home-tab">
-                      <div class="row">
-                        <mfe-loader [config]="rdsDataTableForRoleMfeConfig"></mfe-loader>
-                      </div>
+                    <div class="row">
+                      <mfe-loader [config]="rdsDataTableForMemberMfeConfig"></mfe-loader>
                     </div>
                   </div>
+                  <div class="tab-pane fade" id="Roles" role="tabpanel" aria-labelledby="nav-home-tab">
+                    <div class="row">
+                      <mfe-loader [config]="rdsDataTableForRoleMfeConfig"></mfe-loader>
+                    </div>
+                  </div>
+                </div>
               </div>
             </div>
           </div>
