+++ raaghu-mfe/projects/settings/src/app/app.component.ts
@@ -41,7 +41,6 @@ import { transition,trigger,query,style,  animate,} from '@angular/animations';
 export class AppComponent extends MfeBaseComponent implements OnInit {
   isAnimation: boolean = true;
 
-  isAnimation: boolean = true;
   currentAlerts: any = [];
   editShimmer:boolean=false;
   public rdsAlertMfeConfig: ComponentLoaderOptions = {
